open command now does not advance index
We now keep the index at the position of the file opened, because often after viewing a
file you may want to take action on it, e.g. delete or filter based on it. Previously
default open just advanced the index afterwards. Now it occurs before opening, but only
if a file has not been viewed yet.