covert over existing commands to action module
This is a bit of an interim solution here.  There are a number of uses of Bind to
ad-hoc destination structs, which is working pretty well.  In some cases it's a little
verbose, but we'll see how it goes while we expand out the key binding and the more
specific message setup.