add command package
Ok this undoes the last commit a bit, it just didn't feel constructive to have all these
random packages that mostly are used together.  Let's just have a single package, even if
we can draw some lines between it's components.

This commit also introduces the command.Config and it's Resolve method, which will be key
in making our models easier to understand.