restructured the way tab functions work
I'm setting out to try and solve here the issue of return path for data loads going
to the wrong tab.  In order to do that, we need the stash service to be able to know
which tab a request should be routed to.

This commit lays the groundwork for that, but making the method for creating new tabs
more in line with how commands are now working in each tab model.

Mostly a refactor in preparation for fixing this issue.